src/server/requirements.txt
dataset
click
tracery
stringcase
fastapi
aiofiles
python-multipart
uvicorn[standard]
jinja2
readerwriterlock
requests
hsluv
pillow
banana_dev